User commit

+++ src/main/java/com/example/demo/services/StudentServiceImpl.java
@@ -8,7 +8,7 @@ import com.example.demo.entity.Student;
 @Service
 public class StudentServiceImpl implements StudentService {
 
-    private final Map<Long, Student> store = new HashMap<>();
+    private final Map<Integer, Student> store = new HashMap<>();
     private long counter = 1;
 
     @Override
